Use validation in Controllers

+++ Controllers/CommentController.cs
@@ -24,13 +24,20 @@ namespace api.Controllers
         [HttpGet]
         public async Task<ActionResult> GetAll()
         {
+            var isValid = await _commentRepo.IsCommentValid(ModelState);
+
+            if (isValid != null)
+            {
+                return BadRequest(isValid);
+            }
+
             var comments =  await _commentRepo.GetAllCommentsAsync();
 
             var commentsDTO = comments.Select(x => x.ToCommentDTO());
 
             return Ok(commentsDTO);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<ActionResult> GetById([FromRoute] int id)
         {
             var comment = await _commentRepo.GetCommentByIdAsync(id);
@@ -42,9 +49,16 @@ namespace api.Controllers
             return Ok(comment.ToCommentDTO());
         }
 
-        [HttpPost("{stockId}")]
-        public async Task<ActionResult> Create([FromRoute] int stockId, CreateCommentRequestDTO form)
-        {
+        [HttpPost("{stockId:int}")]
+        public async Task<ActionResult?> Create([FromRoute] int stockId, CreateCommentRequestDTO form)
+        {   
+            var isValid = await _commentRepo.IsCommentValid(ModelState);
+
+            if (isValid != null)
+            {
+                return BadRequest(isValid);
+            }
+
             if (!await _stockRepo.StockExistAsync(stockId))
             {
                 return BadRequest("Stock doesn't exist");
@@ -56,10 +70,17 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id}, commentModel.ToCommentDTO());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
 
         public async Task<ActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDTO updateComment)
         {
+            var isValid = await _commentRepo.IsCommentValid(ModelState);
+
+            if (isValid != null)
+            {
+                return BadRequest(isValid);
+            }
+
             var comment = await _commentRepo.UpdateCommentAsync(id, updateComment);
             if (comment == null)
             {
@@ -68,10 +89,17 @@ namespace api.Controllers
             return Ok(comment.ToCommentDTO());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
 
         public async Task<ActionResult> Delete([FromRoute] int id)
         {
+            var isValid = await _commentRepo.IsCommentValid(ModelState);
+
+            if (isValid != null)
+            {
+                return BadRequest(isValid);
+            }
+
             var comment = await _commentRepo.DeleteCommentAsync(id);
             if (comment == null)
             {
@@ -79,7 +107,5 @@ namespace api.Controllers
             }
             return NoContent();
         }
-
-
     }
 }
